+++ lib/rails-bootstrap-helpers.rb
@@ -4,6 +4,7 @@ module RailsBootstrapHelpers
   module Renderers
     autoload :AbstractButtonRenderer, "rails-bootstrap-helpers/renderers/abstract_button_renderer"
     autoload :AccordionRenderer, "rails-bootstrap-helpers/renderers/accordion_renderer"
+    autoload :ActionLinkRenderer, "rails-bootstrap-helpers/renderers/action_link_renderer"
     autoload :ButtonRenderer, "rails-bootstrap-helpers/renderers/button_renderer"
     autoload :ContentTagRenderer, "rails-bootstrap-helpers/renderers/content_tag_renderer"
     autoload :Renderer, "rails-bootstrap-helpers/renderers/renderer"
@@ -11,6 +12,7 @@ module RailsBootstrapHelpers
 
   module Helpers
     autoload :AccordionHelper, "rails-bootstrap-helpers/helpers/accordion_helper"
+    autoload :ActionLinkHelper, "rails-bootstrap-helpers/helpers/action_link_helper"
     autoload :AlertHelper, "rails-bootstrap-helpers/helpers/alert_helper"
     autoload :BaseHelper, "rails-bootstrap-helpers/helpers/base_helper"
     autoload :ButtonHelper, "rails-bootstrap-helpers/helpers/button_helper"