+++ lib/rails-bootstrap-helpers/renderers/action_link_renderer.rb
@@ -1,0 +1,19 @@
+module RailsBootstrapHelpers::Renderers
+  class ActionLinkRenderer < AbstractButtonRenderer
+    def initialize (template, *args, &block)
+      super template, :link, *args, &block
+    end
+
+    def render
+      append_class "act"
+
+      if style = has_option?("style")
+        unless style.to_s == "default"
+          append_class "act-" + style.to_s
+        end
+      end
+
+      super
+    end
+  end
+end