+++ changelog.md
@@ -11,6 +11,7 @@
 * Add new options helper `append_class!`
 * Add new button helper `button_group`
 * Add support for tooltips to button helpers
+* Add new helper `action_link_to`
 
 ### Bugs Fixed
 * "class" option not being passed through button helpers